Obj-C and Swift bindings update for geometry module.
