+++ modules/geometry/misc/objc/gen_dict.json
@@ -1,5 +1,30 @@
 {
     "namespaces_dict": {
         "cv.fisheye": "fisheye"
+    },
+    "func_arg_fix" : {
+        "Geometry" : {
+            "minEnclosingCircle" : { "points" : {"ctype" : "vector_Point2f"} },
+            "fitEllipse" : { "points" : {"ctype" : "vector_Point2f"} },
+            "approxPolyDP" : { "curve"       : {"ctype" : "vector_Point2f"},
+                               "approxCurve" : {"ctype" : "vector_Point2f"} },
+            "arcLength" : { "curve" : {"ctype" : "vector_Point2f"} },
+            "pointPolygonTest" : { "contour" : {"ctype" : "vector_Point2f"} },
+            "minAreaRect" : { "points" : {"ctype" : "vector_Point2f"} },
+            "getAffineTransform" : { "src" : {"ctype" : "vector_Point2f"},
+                                     "dst" : {"ctype" : "vector_Point2f"} },
+            "convexityDefects" : { "contour"          : {"ctype" : "vector_Point"},
+                                   "convexhull"       : {"ctype" : "vector_int"},
+                                   "convexityDefects" : {"ctype" : "vector_Vec4i"} },
+            "isContourConvex" : { "contour" : {"ctype" : "vector_Point"} },
+            "convexHull" : { "points" : {"ctype" : "vector_Point"},
+                             "hull"   : {"ctype" : "vector_int"},
+                             "returnPoints" : {"ctype" : ""} },
+            "matchShapes" : { "method" : {"ctype" : "ShapeMatchModes"}},
+            "fitLine" : { "distType" : {"ctype" : "DistanceTypes"}}
+        },
+        "Subdiv2D" : {
+            "(void)insert:(NSArray<Point2f*>*)ptvec" : { "insert" : {"name" : "insertVector"} }
+        }
     }
 }
