+++ modules/imgproc/misc/objc/gen_dict.json
@@ -45,39 +45,22 @@
     },
     "func_arg_fix" : {
         "Imgproc" : {
-            "minEnclosingCircle" : { "points" : {"ctype" : "vector_Point2f"} },
-            "fitEllipse" : { "points" : {"ctype" : "vector_Point2f"} },
             "fillPoly"   : { "pts" : {"ctype" : "vector_vector_Point"},
                              "lineType" : {"ctype" : "LineTypes"}},
             "polylines"  : { "pts" : {"ctype" : "vector_vector_Point"},
                              "lineType" : {"ctype" : "LineTypes"} },
             "fillConvexPoly" : { "points" : {"ctype" : "vector_Point"},
                                  "lineType" : {"ctype" : "LineTypes"} },
-            "approxPolyDP" : { "curve"       : {"ctype" : "vector_Point2f"},
-                               "approxCurve" : {"ctype" : "vector_Point2f"} },
-            "arcLength" : { "curve" : {"ctype" : "vector_Point2f"} },
-            "pointPolygonTest" : { "contour" : {"ctype" : "vector_Point2f"} },
-            "minAreaRect" : { "points" : {"ctype" : "vector_Point2f"} },
-            "getAffineTransform" : { "src" : {"ctype" : "vector_Point2f"},
-                                     "dst" : {"ctype" : "vector_Point2f"} },
             "drawContours" : { "contours" : {"ctype" : "vector_vector_Point"},
                                "lineType" : {"ctype" : "LineTypes"} },
             "findContours" : { "contours" : {"ctype" : "vector_vector_Point"},
                                "mode" : {"ctype" : "RetrievalModes"},
                                "method" : {"ctype" : "ContourApproximationModes"} },
-            "convexityDefects" : { "contour"          : {"ctype" : "vector_Point"},
-                                   "convexhull"       : {"ctype" : "vector_int"},
-                                   "convexityDefects" : {"ctype" : "vector_Vec4i"} },
-            "isContourConvex" : { "contour" : {"ctype" : "vector_Point"} },
-            "convexHull" : { "points" : {"ctype" : "vector_Point"},
-                             "hull"   : {"ctype" : "vector_int"},
-                             "returnPoints" : {"ctype" : ""} },
             "getStructuringElement" : { "shape" : {"ctype" : "MorphShapes"} },
             "EMD" : {"lowerBound" : {"defval" : "cv::Ptr<float>()"},
                      "distType" : {"ctype" : "DistanceTypes"}},
             "createLineSegmentDetector" : { "_refine" : {"ctype" : "LineSegmentDetectorModes"}},
             "compareHist" : { "method" : {"ctype" : "HistCompMethods"}},
-            "matchShapes" : { "method" : {"ctype" : "ShapeMatchModes"}},
             "threshold" : { "type" : {"ctype" : "ThresholdTypes"}},
             "connectedComponentsWithStatsWithAlgorithm" : { "ccltype" : {"ctype" : "ConnectedComponentsAlgorithmsTypes"}},
             "GaussianBlur" : { "borderType" : {"ctype" : "BorderTypes"}},
@@ -108,7 +91,6 @@
             "ellipse" : { "lineType" : {"ctype" : "LineTypes"}},
             "erode" : { "borderType" : {"ctype" : "BorderTypes"}},
             "filter2D" : { "borderType" : {"ctype" : "BorderTypes"}},
-            "fitLine" : { "distType" : {"ctype" : "DistanceTypes"}},
             "line" : { "lineType" : {"ctype" : "LineTypes"}},
             "matchTemplate" : { "method" : {"ctype" : "TemplateMatchModes"}},
             "morphologyEx" : { "op" : {"ctype" : "MorphTypes"},
@@ -126,9 +108,6 @@
             "warpAffine" : { "borderMode": {"ctype" : "BorderTypes"}},
             "warpPerspective" : { "borderMode": {"ctype" : "BorderTypes"}},
             "getTextSize" : { "fontFace": {"ctype" : "HersheyFonts"}}
-        },
-        "Subdiv2D" : {
-            "(void)insert:(NSArray<Point2f*>*)ptvec" : { "insert" : {"name" : "insertVector"} }
         }
     },
     "type_dict": {
