+++ modules/imgproc/misc/objc/test/ImgprocTest.swift
@@ -122,26 +122,6 @@ class ImgprocTest: OpenCVTestCase {
         XCTAssertEqual(src.rows(), Core.countNonZero(src: dst))
     }
 
-    func testApproxPolyDP() {
-        let curve = [Point2f(x: 1, y: 3), Point2f(x: 2, y: 4), Point2f(x: 3, y: 5), Point2f(x: 4, y: 4), Point2f(x: 5, y: 3)]
-
-        var approxCurve = [Point2f]()
-
-        Imgproc.approxPolyDP(curve: curve, approxCurve: &approxCurve, epsilon: OpenCVTestCase.EPS, closed: true)
-
-        let approxCurveGold = [Point2f(x: 1, y: 3), Point2f(x: 3, y: 5), Point2f(x: 5, y: 3)]
-
-        XCTAssert(approxCurve == approxCurveGold)
-    }
-
-    func testArcLength() {
-        let curve = [Point2f(x: 1, y: 3), Point2f(x: 2, y: 4), Point2f(x: 3, y: 5), Point2f(x: 4, y: 4), Point2f(x: 5, y: 3)]
-
-        let arcLength = Imgproc.arcLength(curve: curve, closed: false)
-
-        XCTAssertEqual(5.656854249, arcLength, accuracy:0.000001)
-    }
-
     func testBilateralFilterMatMatIntDoubleDouble() throws {
         Imgproc.bilateralFilter(src: gray255, dst: dst, d: 5, sigmaColor: 10, sigmaSpace: 5)
 
@@ -315,24 +295,6 @@ class ImgprocTest: OpenCVTestCase {
         XCTAssertEqual(1.0, distance, accuracy: OpenCVTestCase.EPS)
     }
 
-    func testContourAreaMat() throws {
-        let contour = Mat(rows: 1, cols: 4, type: CvType.CV_32FC2)
-        try contour.put(row: 0, col: 0, data: [0, 0, 10, 0, 10, 10, 5, 4] as [Float])
-
-        let area = Imgproc.contourArea(contour: contour)
-
-        XCTAssertEqual(45.0, area, accuracy: OpenCVTestCase.EPS)
-    }
-
-    func testContourAreaMatBoolean() throws {
-        let contour = Mat(rows: 1, cols: 4, type: CvType.CV_32FC2)
-        try contour.put(row: 0, col: 0, data: [0, 0, 10, 0, 10, 10, 5, 4] as [Float])
-
-        let area = Imgproc.contourArea(contour: contour, oriented: true)
-
-        XCTAssertEqual(45.0, area, accuracy: OpenCVTestCase.EPS)
-    }
-
     func testConvertMapsMatMatMatMatInt() throws {
         let map1 = Mat(rows: 1, cols: 4, type: CvType.CV_32FC1, scalar: Scalar(1))
         let map2 = Mat(rows: 1, cols: 4, type: CvType.CV_32FC1, scalar: Scalar(2))
@@ -377,38 +339,6 @@ class ImgprocTest: OpenCVTestCase {
         Imgproc.convexHull(points: points, hull: &hull)
 
         XCTAssert([0, 1, 2, 3] == hull)
-    }
-
-    func testConvexHullMatMatBooleanBoolean() {
-        let points = [Point(x: 2, y: 0),
-                      Point(x: 4, y: 0),
-                      Point(x: 3, y: 2),
-                      Point(x: 0, y: 2),
-                      Point(x: 2, y: 1),
-                      Point(x: 3, y: 1)]
-
-        var hull = [Int32]()
-
-        Imgproc.convexHull(points: points, hull: &hull, clockwise: true)
-
-        XCTAssert([3, 2, 1, 0] == hull)
-    }
-
-    func testConvexityDefects() throws {
-        let points = [Point(x: 20, y: 0),
-                      Point(x: 40, y: 0),
-                      Point(x: 30, y: 20),
-                      Point(x: 0,  y: 20),
-                      Point(x: 20, y: 10),
-                      Point(x: 30, y: 10)]
-
-        var hull = [Int32]()
-        Imgproc.convexHull(points: points, hull: &hull)
-
-        var convexityDefects = [Int4]()
-        Imgproc.convexityDefects(contour: points, convexhull: hull, convexityDefects: &convexityDefects)
-
-        XCTAssertTrue(Int4(v0: 3, v1: 0, v2: 5, v3: 3620) == convexityDefects[0])
     }
 
     func testCornerEigenValsAndVecsMatMatIntInt() throws {
@@ -733,19 +663,6 @@ class ImgprocTest: OpenCVTestCase {
         try assertMatEqual(gray2, dst)
     }
 
-    func testGetAffineTransform() throws {
-        let src = [Point2f(x: 2, y: 3), Point2f(x: 3, y: 1), Point2f(x: 1, y: 4)]
-        let dst = [Point2f(x: 3, y: 3), Point2f(x: 7, y: 4), Point2f(x: 5, y: 6)]
-
-        let transform = Imgproc.getAffineTransform(src: src, dst: dst)
-
-        let truth = Mat(rows: 2, cols: 3, type: CvType.CV_64FC1)
-
-        try truth.put(row: 0, col: 0, data: [-8.0, -6.0, 37.0])
-        try truth.put(row: 1, col: 0, data: [-7.0, -4.0, 29.0])
-        try assertMatEqual(truth, transform, OpenCVTestCase.EPS)
-    }
-
     func testGetDerivKernelsMatMatIntIntInt() throws {
         let kx = Mat(rows: imgprocSz, cols: imgprocSz, type: CvType.CV_32F)
         let ky = Mat(rows: imgprocSz, cols: imgprocSz, type: CvType.CV_32F)
@@ -816,18 +733,6 @@ class ImgprocTest: OpenCVTestCase {
         Imgproc.getRectSubPix(image: src, patchSize: patchSize, center: center, patch: dst)
 
         truth = Mat(rows: 5, cols: 5, type: CvType.CV_32F, scalar: Scalar(2))
-        try assertMatEqual(truth!, dst, OpenCVTestCase.EPS)
-    }
-
-    func testGetRotationMatrix2D() throws {
-        let center = Point2f(x: 0, y: 0)
-
-        dst = Imgproc.getRotationMatrix2D(center: center, angle: 0, scale: 1)
-
-        truth = Mat(rows: 2, cols: 3, type: CvType.CV_64F)
-        try truth!.put(row: 0, col: 0, data: [1.0, 0.0, 0.0])
-        try truth!.put(row: 1, col: 0, data: [0.0, 1.0, 0.0])
-
         try assertMatEqual(truth!, dst, OpenCVTestCase.EPS)
     }
 
@@ -1021,25 +926,6 @@ class ImgprocTest: OpenCVTestCase {
         try assertMatEqual(truth!, dst, OpenCVTestCase.EPS)
     }
 
-    func testInvertAffineTransform() throws {
-        let src = Mat(rows: 2, cols: 3, type: CvType.CV_64F, scalar: Scalar(1))
-
-        Imgproc.invertAffineTransform(M: src, iM: dst)
-
-        truth = Mat(rows: 2, cols: 3, type: CvType.CV_64F, scalar: Scalar(0))
-        try assertMatEqual(truth!, dst, OpenCVTestCase.EPS)
-    }
-
-    func testIsContourConvex() {
-        let contour1 = [Point(x: 0, y: 0), Point(x: 10, y: 0), Point(x: 10, y: 10), Point(x: 5, y: 4)]
-
-        XCTAssertFalse(Imgproc.isContourConvex(contour: contour1))
-
-        let contour2 = [Point(x: 0, y: 0), Point(x: 10, y: 0), Point(x: 10, y: 10), Point(x: 5, y: 6)]
-
-        XCTAssert(Imgproc.isContourConvex(contour: contour2))
-    }
-
     func testLaplacianMatMatInt() throws {
         Imgproc.Laplacian(src: gray0, dst: dst, ddepth: CvType.CV_8U)
 
@@ -1103,27 +989,6 @@ class ImgprocTest: OpenCVTestCase {
         // TODO_: write better test
     }
 
-    func testMinAreaRect() {
-        let points = [Point2f(x: 1, y: 1), Point2f(x: 5, y: 1), Point2f(x: 4, y: 3), Point2f(x: 6, y: 2)]
-
-        let rrect = Imgproc.minAreaRect(points: points)
-
-        XCTAssertEqual(Size2f(width: 5, height: 2), rrect.size)
-        XCTAssertEqual(0.0, rrect.angle)
-        XCTAssertEqual(Point2f(x: 3.5, y: 2), rrect.center)
-    }
-
-    func testMinEnclosingCircle() {
-        let points = [Point2f(x: 0, y: 0), Point2f(x: -100, y: 0), Point2f(x: 0, y: -100), Point2f(x: 100, y: 0), Point2f(x: 0, y: 100)]
-        let actualCenter = Point2f()
-        var radius:Float = 0
-
-        Imgproc.minEnclosingCircle(points: points, center: actualCenter, radius: &radius)
-
-        XCTAssertEqual(Point2f(x: 0, y: 0), actualCenter)
-        XCTAssertEqual(100.0, radius, accuracy: 1.0)
-    }
-
     func testMorphologyExMatMatIntMat() throws {
         Imgproc.morphologyEx(src: gray255, dst: dst, op: MorphTypes.MORPH_GRADIENT, kernel: gray0)
 
@@ -1157,15 +1022,6 @@ class ImgprocTest: OpenCVTestCase {
         try truth!.put(row: 0, col: 0, data: [1, 0] as [Int8])
         try truth!.put(row: 1, col: 0, data: [1, 0] as [Int8])
         try assertMatEqual(truth!, dst)
-    }
-
-    func testPointPolygonTest() {
-        let contour = [Point2f(x: 0, y: 0), Point2f(x: 1, y: 3), Point2f(x: 3, y: 4), Point2f(x: 4, y: 3), Point2f(x: 2, y: 1)]
-        let sign1 = Imgproc.pointPolygonTest(contour: contour, pt: Point2f(x: 2, y: 2), measureDist: false)
-        XCTAssertEqual(1.0, sign1)
-
-        let sign2 = Imgproc.pointPolygonTest(contour: contour, pt: Point2f(x: 4, y: 4), measureDist: true)
-        XCTAssertEqual(-sqrt(0.5), sign2)
     }
 
     func testPreCornerDetectMatMatInt() throws {
